chore: extracts sample_data
